Separate webhook and zabbix applications
In order to separate webhook and zabbix applications, we separate the
Dockerfile and requirements.txt. We are now able to build two separate
Docker images and run them separately.

In this same change set, we optimized the Docker images building. For
this, we used multistage building and lightweight base images (it is now
based on the `python:3.6-alpine` base image). We managed to reduce from
1.1GB to around 90MB the image sizes.