Handle invalid JSON gracefully
It now responds with:

* 400 Bad Request when data sent contains in its body an invalid JSON.
* 200 Success when data sent is a valid JSON.

In both cases, we also responds with a descriptive message in JSON in
the response body.

Improve way webhook data is handled

Increase space for module name in log message

Save WIP

Change for new format of webhook event

It now consists of two distincts fields `event` and `timestamp`.
The `event` field is a JSON and `timestamp` is a POSIX timestamp.
The POST data containining these fields has Content-Type
application/x-www-form-urlencoded.