Update Makefile and add urllib3 to logging
In Makefile, `IMAGE_NAME`, `IMAGE_VERSION` and `LOCAL_TAG` could not be
set externally (via environment variable) when running the Makefile. By
adding `?=` in its definition, we allow it to be set externally.

The library `urllib3` was included to logging and its `propagate`
attribute made false so its log does not appear in logs.