Change playback to be an upsert
Playback in recordings table were a complete replacement. Instead, we
should upsert new event playback. If recording's playbacks already have
a playback with the same format as the new event playback, we simply replace it.
OTOH if the new event playback format is not in recording's playbacks yet, we
should append it.